we finna go c++26 yalls

+++ common/include/kernel/elf64_dynamic_exec.hpp
@@ -17,7 +17,7 @@ protected:
     virtual bool load_syms() override;
     virtual void process_headers() override;
 public:
-    virtual addr_t resolve(uint64_t offs) const override;    
+    virtual addr_t resolve(uintptr_t offs) const override;    
     virtual addr_t resolve(elf64_sym const& sym) const override;
     virtual ~elf64_dynamic_executable();
     elf64_dynamic_executable(file_node* n, size_t stack_sz = S04, size_t tls_sz = S04, uintptr_t base_offset = 0UL);
